backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=0.177.0",
    "crewai-tools>=0.69.0",
    "fastapi>=0.116.1",
    "langchain-core>=0.3.75",
    "langchain-mistralai>=0.2.11",
    "python-dotenv>=1.1.1",
    "qdrant-client>=1.15.1",
    "uvicorn>=0.35.0",
]
